ring: Fix image memory leak
... and various code correctness and style issues.

Use unique_ptr instead of naked new/delete, and use a std::multimap for
m_image_by_top_map.

https://gitlab.gnome.org/GNOME/vte/-/issues/255

+++ src/image.hh
@@ -31,7 +31,7 @@ private:
         vte::Freeable<cairo_surface_t> m_surface{};
 
         // Draw/prune priority, must be unique
-        int m_priority;
+        size_t m_priority;
 
         // Image dimensions in pixels
         int m_width_pixels;
@@ -47,7 +47,7 @@ private:
 
 public:
         Image(vte::Freeable<cairo_surface_t> surface,
-              int priority,
+              size_t priority,
               int width_pixels,
               int height_pixels,
               int col,
